whole model basic exceptions are done

+++ src/e-commerce/product/product.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
@@ -7,8 +7,10 @@ import { UpdateProductDto } from './dto/update-product.dto';
 export class ProductService {
 	constructor(private prisma:PrismaService){}
 
-	create(body:CreateProductDto){
-		return this.prisma.product.create({data: body});
+	create(data:CreateProductDto){
+		if(data.price < 0) throw new BadRequestException('Price not be negative')
+		if(data.stock < 0) throw new BadRequestException('Stock not be negative')	
+		return this.prisma.product.create({data: data});
 	}
 	getAllProducts(){
 		return this.prisma.product.findMany({
@@ -17,8 +19,8 @@ export class ProductService {
 			}
 		});
 	}
-	getProductById(id:number){
-		return this.prisma.product.findUnique({
+	async getProductById(id:number){
+		const product = await this.prisma.product.findUnique({
 			where:{
 				id
 			},
@@ -30,20 +32,29 @@ export class ProductService {
 				}
 			}
 		})
+		if(!product) throw new NotFoundException('Product not found');
+		return product;
 	}
-	updateProductById(id:number,body:UpdateProductDto){
+	async updateProductById(id:number,data:UpdateProductDto){
+		const product = await this.prisma.product.findUnique({where:{id}})
+		if(!product) throw new NotFoundException('Product not found');
+		if(data.price !== undefined && data.price < 0) throw new BadRequestException('Price not be negative')
+		if(data.stock !== undefined && data.stock < 0) throw new BadRequestException('Stock not be negative')
 		return this.prisma.product.update({
 			where:{
 				id
 			},
-			data:body
+			data:data
 		})
 	}
-	deleteProductById(id:number){
+	async deleteProductById(id:number){
+		const product = await this.prisma.product.findUnique({where:{id}})
+		if(!product) throw new NotFoundException('Product not found');
 		return this.prisma.product.delete({
 			where:{
 				id
 			}
 		})
 	}
+	
 }
